Update patch set 1
Patch Set 1:

Actually this patch should address a regression that occurred with the introduction of the LiveLockScreenService.

I have the Colors live lock screen selected and when I toggle off the Live lock screen feature, turn the screen off then back on, the live lock screen is gone.  If this is not the case for you, and you are on a build that includes all the latest live lock screen related changes (April 2nd should be good) then please file a bug at https://jira.cyanogenmod.org/browse/NIGHTLIES

Patch-set: 1